main executed in raw-probe
