+++ meet-data/raw/raw-probe
@@ -9,9 +9,15 @@ import datetime
 import os
 import shutil
 import sys
-sys.path.append('../')
+sys.path.append('scripts')
 import probe
 import urllib.request
+
+# URL needs updating every year.
+# Also be careful to check if there are any result links that don't contain
+# the string "results", since they won't get picked up by the script below.
+if datetime.datetime.now().strftime("%Y") != "2017":
+    print("Warning: RAW fetch URL needs updating for new year.", file=sys.stderr)
 
 FEDDIR = os.path.dirname(os.path.realpath(__file__))
 URL = "http://rawpowerlifting.com/2017_results/"
@@ -22,11 +28,6 @@ def color(s):
     return "\033[1;30m" + s + "\033[0;m"
 
 def getmeetlist(html):
-    # URL needs updating every year.
-    # Also be careful to check if there are any result links that don't contain
-    # the string "results", since they won't get picked up by the script below.
-    if datetime.datetime.now().strftime("%Y") != "2017":
-        print("Warning: RAW fetch URL needs updating for new year.", file=sys.stderr)
 
     soup = BeautifulSoup(html, 'html.parser')
 
@@ -49,5 +50,17 @@ def getmeetlist(html):
 
     return urls
 
+def main():
+    html = probe.gethtml(URL)
+    meetlist = getmeetlist(html)
+
+    known = probe.getenteredurls(FEDDIR)
+    unknown = probe.getunknownmeets(meetlist, known)
+
+    for m in unknown:
+        print(color('[RAW] ') + m)
+
+    print(color('[RAW] ') + "Continue working through archive.")
+
 if __name__ == '__main__':
-    probe.main('[RAW]', color, getmeetlist, URL)
+    main()
